move toast pop up to root component to clarify for reusability
